lowestOrder changed to long long from int (64 bit problem) Added post-build events for the creation and move of 64bit executables (with 64-suffix) into bin folder

+++ Autonomic/autonomic.cpp
@@ -2321,7 +2321,7 @@ int AgentBase::_atomicMessageDecide( UUID *id, UUID *q, int orderq, int roundq, 
 }
 
 int AgentBase::_atomicMessageAttemptCommit( UUID queue ) {
-	int lowestOrder;
+	long long lowestOrder; // int lowestOrder;
 	mapAtomicMessage::iterator iM;
 	AtomicMessage *amLow;
